Implement account cloning feature in PmTask (#20)
* Framework for clone account

Sketching out a basic framework for cloning an account

* Make parameters cleaner

* Basic shell of cloning logic

* Print out some useful results

* Reduce code verbosity a bit

Let's add a method to make throw on error simpler

* Add sync for project fields

Project Fields can be synced, but no update is possible, so we need to modify the sync logic slightly.

* Added ResourceSkills

* Basic syncing for resources includes mapping skills/teams

* Finish off all the clone helper lists

* Load in other dependent data for matching

* Fix nullability of guids

* Update and delete resources

* Filter to active resources

* Fix resource comparison

* Working on cloning projects

* Partial implementation for ProjectMembers

* Sync ProjectFieldValues across Projects

* Sync TaskStatuses across Projects

Currently, there is no way to create/update the IsDone property via the API - value is always false

* Match TaskPriorities

* Sync TaskFields across Projects

* Revert Match TaskPriorities

Task Priorities are assigned as integers, so we don't need to map to the actual id

* Clone basic Tasks

Basic Tasks are cloned - still missing some data like files, comments, etc.  Main issue seems to be with nested tasks and ordering - see Hourly Rates project.

* Fix typo around console log

* Process Tasks in correct order (Wbs)

* Indent Subtasks

* Only Indent Subtasks if required

* Clone TaskFields outside Projects

* Clone TaskFieldValues

* Clone TaskTags

* Clean up logging across projects

* Add export of mapping files

* Add cloning for timesheets

* Ensure account is cloned

* Added Update and Delete for TimesheetEntries

* When indenting, only indent subtasks in the same project

---------

Co-authored-by: Waynne Human <waynne@gmail.com>